requirements.txt
streamlit
langchain
langchain-groq
pandas
python-dotenv
pydantic
chromadb
huggingface-hub
langchain-community
sentence-transformers
pdfminer.six
python-docx
fitz
pdfminer.six 